+++ dataworkspace/dataworkspace/tests/requesting_data/test_requesting_data_unit.py
@@ -2,14 +2,16 @@ from http import HTTPStatus
 from unittest.mock import patch
 from unittest import TestCase
 
+from bs4 import BeautifulSoup
 from django.contrib.auth import get_user_model
 from django.test import Client
 from django.urls import reverse
 
+import mock
 import pytest
 
 from dataworkspace.tests import factories
-from dataworkspace.tests.common import get_http_sso_data
+from dataworkspace.tests.common import BaseTestCase, get_http_sso_data
 from dataworkspace.apps.datasets.models import RequestingDataset, SensitivityType
 from dataworkspace.apps.datasets.requesting_data.forms import (
     DatasetDescriptionsForm,
@@ -376,3 +378,33 @@ class RequestingDataDeleteTestCase(TestCase):
             reverse("delete-requesting-dataset-journey", args=[requesting_dataset.id])
         )
         assert response.status_code == 302
+
+
+@pytest.mark.django_db
+class TestTrackerViewSubmission(TestCase):
+    @mock.patch("dataworkspace.apps.datasets.requesting_data.views.create_support_request")
+    def test_create_tagged_support_request(self, mock_create_request):
+        mock_create_request.return_value = 999
+        self.user = factories.UserFactory.create(is_superuser=False)
+        self.client = Client(**get_http_sso_data(self.user))
+        self.requesting_dataset = factories.RequestingDataSetFactory.create(user=self.user.id)
+        # pylint: disable=attribute-defined-outside-init
+        response = self.client.post(
+            reverse("requesting-data-tracker", args={(self.requesting_dataset.id)}),
+            data={
+                "requesting_dataset": self.requesting_dataset.id,
+                "user": self.user,
+                "email": self.user.email,
+                "message": "A new dataset has been requested.",
+                "tag": "data_request",
+                "requester": self.user,
+            },
+        )
+        assert "/requesting-data/submission/999" in response.url
+
+        mock_create_request.assert_called_once_with(
+            user=mock.ANY,
+            email=self.user.email,
+            message="A new dataset has been requested.",
+            tag="data_request",
+        )
